tests/wpt/tests/html/canvas/element/canvas-host/2d.canvas.host.initial.reset.path.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.canvas.host.initial.reset.path</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.canvas.host.initial.reset.path</h1>
<p class="desc">Resetting the canvas state resets the current path</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Resetting the canvas state resets the current path");
_addTest(function(canvas, ctx) {

  canvas.width = 100;
  ctx.rect(0, 0, 100, 50);
  canvas.width = 100;
  ctx.fillStyle = '#f00';
  ctx.fill();
  _assertPixel(canvas, 20,20, 0,0,0,0);

});
</script>
